requirements.txt
google-generativeai
moviepy==1.0.3
requests>=2.31.0
python-dotenv>=1.0.0
yt-dlp